Create ups_autochecker.py
Checking on our Shipments excel for every package that needs to be checked, POST a request to UPS's API then get the response ( json ) in a file ( getstatus.json ). The file then gets parsed and we get the data that we need in it. Final step is to change color and text in our Shipment file in excel automatically.